Review fixes
Moved and renamed File to CodeFile.
