+++ mentat/app.py
@@ -1,4 +1,5 @@
 import argparse
+import glob
 import logging
 from typing import Iterable, Optional
 
@@ -6,6 +7,7 @@ from termcolor import cprint
 
 from .code_change import CodeChange
 from .code_change_display import print_change
+from .code_file import parse_intervals
 from .code_file_manager import CodeFileManager
 from .config_manager import ConfigManager, mentat_dir_path
 from .conversation import Conversation
@@ -36,7 +38,34 @@ def run_cli():
     args = parser.parse_args()
     paths = args.paths
     exclude_paths = args.exclude
-    run(paths, exclude_paths)
+    # Expanding paths as soon as possible because some shells such as zsh automatically
+    # expand globs and we want to avoid differences in functionality between shells
+    run(expand_paths(paths), expand_paths(exclude_paths))
+
+
+def expand_paths(paths: Iterable[str]) -> Iterable[str]:
+    globbed_paths = set()
+    invalid_paths = []
+    for path in paths:
+        new_paths = glob.glob(pathname=path, recursive=True)
+        if new_paths:
+            globbed_paths.update(new_paths)
+        else:
+            split = path.rsplit(":", 1)
+            p = split[0]
+            intervals = parse_intervals(split[1])
+            if Path(p).exists() and intervals is not None:
+                globbed_paths.add(path)
+            else:
+                invalid_paths.append(path)
+    if invalid_paths:
+        cprint(
+            "The following paths do not exist:",
+            "light_yellow",
+        )
+        print("\n".join(invalid_paths))
+        exit()
+    return globbed_paths
 
 
 def run(paths: Iterable[str], exclude_paths: Optional[Iterable[str]] = None):
